uigen: split code into separate source and header
This adds a --header option to uigen to inform it that it's
output will be a header rather than a source file.

Add a --header-guard option to specify the #include guard
in the generated header.

Add menu language syntax to optionally place boilerplate
text / includes in only the header or source output.

Extend default template to vary output accordingly.

Add Makefile rule to ensure the header is generated.